remove 'is_fragment' from genedata and have it in 'gene' as it can change through time depending on analysis
